requirements.txt
Flask==2.0.1
google-cloud-secret-manager
mysql-connector-python==8.0.25